こうすると、1が取り出せるみたい。 remove(Object)が優先されるみたい。boxingしているかの違いが問題なのかな？

+++ myLib.MIT/myLib.container.mkv/src/test/java/com/ttProject/container/mkv/test/ListTest.java
@@ -26,7 +26,7 @@ public class ListTest {
 		list.add(5);
 		list.add(2);
 		logger.info(list);
-		logger.info(list.remove(1));
+		logger.info(list.remove((Integer)1));
 		logger.info(list);
 	}
 }
